+++ server/src/nest/trip-invite/trip-invite.module.ts
@@ -1,13 +1,13 @@
+import { RateLimitModule } from '../common/rate-limit.module';
 import { Module } from '@nestjs/common';
 import { TripInviteLinkController, TripInviteController } from './trip-invite.controller';
 import { TripInviteService } from './trip-invite.service';
-import { RateLimitService } from '../auth/rate-limit.service';
 import { PermissionsModule } from '../permissions/permissions.module';
 import { AuditModule } from '../audit/audit.module';
 
 @Module({
-  imports: [PermissionsModule, AuditModule],
+  imports: [RateLimitModule, PermissionsModule, AuditModule],
   controllers: [TripInviteLinkController, TripInviteController],
-  providers: [TripInviteService, RateLimitService],
+  providers: [TripInviteService],
 })
 export class TripInviteModule {}
